web_interface/backend/requirements.txt
fastapi>=0.89.0
uvicorn>=0.20.0
pydantic>=1.10.4
jinja2>=3.1.2
python-multipart>=0.0.5
aiofiles>=0.8.0
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1
websockets>=10.4
SpeechRecognition>=3.9.0
python-dotenv>=0.21.0
itsdangerous>=2.1.2
httpx>=0.23.0 